+++ _config/shortcodes/async/asyncShortcodes.js
@@ -1,0 +1,11 @@
+import {figureGroup} from "./figureGroup.js";
+import {imageBg} from "./imageBg.js";
+import {image} from "./image.js";
+import {showcase} from "./showcase.js";
+
+export default {
+    figureGroup,
+    imageBg,
+    image,
+    showcase
+};
