+++ _config/shortcodes/shortcodes.js
@@ -1,0 +1,9 @@
+import {listItems} from './listItems.js';
+import {primaryNavigation} from './primaryNavigation.js';
+import {secondaryNavigation} from './secondaryNavigation.js';
+
+export default {
+    listItems,
+    primaryNavigation,
+    secondaryNavigation
+};
