+++ src/_data/dev.js
@@ -1,4 +1,4 @@
-const fs = require('fs');
+import fs from 'fs';
 
 const visualizationsData = fs.readFileSync(`${process.cwd()}/src/_data/visualizations_list.json`, 'utf-8')
 const visualizations = JSON.parse(visualizationsData);
@@ -6,7 +6,7 @@ const visualizations = JSON.parse(visualizationsData);
 const websData = fs.readFileSync(`${process.cwd()}/src/_data/websites_list.json`, 'utf-8')
 const websites = JSON.parse(websData);
 
-
-const dev={"visualizations":visualizations, "websites":websites};
-
-module.exports = dev;
+export default {
+    "visualizations":visualizations, 
+    "websites":websites
+};
